django-api/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
PyJWT==2.8.0
cryptography==41.0.7
python-jose[cryptography]==3.3.0
requests==2.31.0
django-cors-headers==4.3.1